fix output 2

+++ commands/fun/ytp.js
@@ -125,14 +125,17 @@ class ytpCommand extends Command {
 			
 			if (url) {
 				return youtubedl.exec(url, ['--max-filesize', '50m', '--format=mp4', '-o', `./asset/ytp/userVid/${message.id}.mp4`], {}, function(err, output) {
+					console.log(output);
 					if (err) {
 						console.error(err);
 						loadingmsg.delete();
 						return message.channel.send('An error has occured, I can\'t download from the link you provided. Is it an mp4?');
 					} else {
-						if (output[2].includes('File is larger than max-filesize')) {
-							loadingmsg.delete();
-							return message.channel.send(output[2]);
+						if (output[2]) {
+							if (output[2].includes('File is larger than max-filesize')) {
+								loadingmsg.delete();
+								return message.channel.send(output[2]);
+							}
 						} 
 
 						let mp4 = [];
